agent/pyproject.toml
[tool.poetry]
name = "vargasjr-dev-agent"
version = "0.0.10"
description = "Vargas Jr Dev Agent"
authors = []
keywords = []
license = "MIT"
classifiers = []
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.12"
vellum-ai = "0.13.0"
sqlmodel = "^0.0.22"
psycopg = {extras = ["binary"], version = "^3.2.3"}
boto3 = "^1.35.91"
google-api-python-client = "^2.157.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"

[tool.poetry.scripts]
agent = "src.cli:main"


[tool.poetry.group.dev.dependencies]
mypy = "1.11.0"
pytest = "^7.4.0"
pytest-mock = "3.10.0"
python-dotenv = "^1.0.1"
pytest-postgresql = "^6.1.1"
ipdb = "^0.13.13"

[tool.mypy]
plugins = ["pydantic.mypy", "vellum.plugins.vellum_mypy"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:pytz",
    "ignore:Load is deprecated on a client fixture:DeprecationWarning",
]
